Module Data_encoding__Uint_option

Optional non-negative integers can be represented by an int provided that None is encoded as -1 and Some x is encoded by x. This module provides a type abstraction to maintain this invariant.

This optimized representation avoids the allocation of a block in the heap and also removes one indirection when reading such an optional non-negative integer.

raises [Assert_failure]

when assumptions are broken. This module is meant to be used internally only, covered by an extensive test-suite with assertion checks on, and packaged with assertion checks off.